Move `getFallbackLocale` to the only place where it's used
